move route53 and acm related stuff to the separate files, add tags for route53 record, refactoring
