+++ src/domain-config.ts
@@ -5,11 +5,9 @@
 import * as AWS from "aws-sdk"; // imported for Types
 import DomainInfo = require("./domain-info");
 import Globals from "./globals";
-import {CustomDomain, Route53Params} from "./types";
+import {CustomDomain, Route53Params, Tags} from "./types";
 
 class DomainConfig {
-
-    public acm: any;
 
     public givenDomainName: string;
     public basePath: string | undefined;
@@ -29,12 +27,11 @@ class DomainConfig {
     public autoDomainWaitFor: string | undefined;
     public route53Params: Route53Params;
     public preserveExternalPathMappings: boolean | undefined;
-
+    public tags: Tags[] | undefined;
     public domainInfo: DomainInfo | undefined;
     public apiId: string | undefined;
     public apiMapping: AWS.ApiGatewayV2.GetApiMappingResponse;
     public allowPathMatching: boolean | false;
-    public region: string;
 
     constructor(config: CustomDomain) {
 
@@ -85,15 +82,6 @@ class DomainConfig {
             throw new Error(`${securityPolicyDefault} is not a supported securityPolicy, use tls_1_0 or tls_1_2.`);
         }
         this.securityPolicy = tlsVersionToUse;
-
-        this.region = Globals.defaultRegion;
-        if (this.endpointType === Globals.endpointTypes.regional) {
-            this.region = Globals.serverless.providers.aws.getRegion();
-        }
-        const acmCredentials = Object.assign(
-            {}, Globals.serverless.providers.aws.getCredentials(), {region: this.region}
-        );
-        this.acm = new Globals.serverless.providers.aws.sdk.ACM(acmCredentials);
 
         const defaultRoutingPolicy = Globals.routingPolicies.simple;
         const routingPolicy = config.route53Params?.routingPolicy?.toLowerCase() ?? defaultRoutingPolicy;
